Depreciated CollectionItem and SearchItem to new BookCard component
