+++ .prettierrc.json
@@ -1,0 +1,3 @@
+{
+	"tabWidth": 4
+}
