+++ frontend/src/components/Collection.js
@@ -1,7 +1,8 @@
 import React, { Component } from "react";
-import CollectionItem from "./CollectionItem";
 import PropTypes from "prop-types";
 import { CardDeck } from "react-bootstrap";
+
+import BookCard from "./BookCard";
 
 /*
 The collection contains collection-items, which themselves contain books to be
@@ -9,36 +10,36 @@ displayed.
 This class deals with displaying those collection items.
 */
 class Collection extends Component {
-  displayCollection = () => {
-    const { currentCollection } = this.props;
-    if (
-      currentCollection !== null &&
-      typeof currentCollection.books !== "undefined"
-    ) {
-      return (
-        <CardDeck>
-          {currentCollection.books.map((book) => (
-            <CollectionItem
-              key={book.id}
-              book={book}
-              removeBook={this.props.removeBook}
-              addToCollection={this.props.addToCollection}
-              userCollections={this.props.userCollections}
-              editable={this.props.editable}
-            />
-          ))}
-        </CardDeck>
-      );
-    }
-  };
+    displayCollection = () => {
+        const { currentCollection } = this.props;
+        if (
+            currentCollection !== null &&
+            typeof currentCollection.books !== "undefined"
+        ) {
+            return (
+                <CardDeck>
+                    {currentCollection.books.map((book) => (
+                        <BookCard
+                            key={book.id}
+                            book={book}
+                            removeBook={this.props.removeBook}
+                            addToCollection={this.props.addToCollection}
+                            userCollections={this.props.userCollections}
+                            editable={this.props.editable}
+                        />
+                    ))}
+                </CardDeck>
+            );
+        }
+    };
 
-  render() {
-    return <div>{this.displayCollection()}</div>;
-  }
+    render() {
+        return <div>{this.displayCollection()}</div>;
+    }
 }
 
 Collection.propTypes = {
-  currentCollection: PropTypes.object.isRequired,
+    currentCollection: PropTypes.object.isRequired,
 };
 
 export default Collection;
