+++ frontend/src/components/SearchResults.js
@@ -1,24 +1,23 @@
 import React, { Component } from "react";
 import { CardDeck } from "react-bootstrap";
-
-import SearchResultsItem from "../components/SearchResultsItem.js";
+import BookCard from "../components/BookCard.js";
 
 class SearchResults extends Component {
-  constructor(props) {
-    super(props);
-  }
+    constructor(props) {
+        super(props);
+    }
 
-  render() {
-    return (
-      <div>
-        <CardDeck>
-          {this.props.books.map((book) => (
-            <SearchResultsItem key={book.id} book={book} />
-          ))}
-        </CardDeck>
-      </div>
-    );
-  }
+    render() {
+        return (
+            <div>
+                <CardDeck>
+                    {this.props.books.map((book) => (
+                        <BookCard key={book.id} book={book} editable="false" />
+                    ))}
+                </CardDeck>
+            </div>
+        );
+    }
 }
 
 export default SearchResults;
